Fixed system calls for linux.
Linux uses fastcall calling convention (i.e., parameters are passed in
registers rather than on the stack).  The original calling convention was moved
to another file (system-cdecl.inc), a new file (system-fastcall.inc) was added
that uses the fastcall convention, and system.inc was changed to include
system-fastcall.inc.  At some point this should be changed so that the correct
calling convention is detected according to the architecture.