Added MULX instruction with 64-bit result.
Added an instruction (44 - MULX) to the SECD instruction set that pops two
32-bit integers off the stack, multiplies them, and then pushes the 64-bit
result onto the stack as a cons pair with the low order 32-bits in the car
position and the high order 32-bits in the cdr position.  This will enable
implementation of big integers by allowing multiplication with carry.