Added PEXP instruction & PUTEXP command.
The PEXP instruction prints the expression at the top of the stack.  It does so
without removing it, so that insertion of this instruction does not affect
program execution.

The PUTEXP command invokes the PEXP instruction.  It may be wrapped around an
expression without changing its value -- i.e., the presence of the PUTEXP
command is transparent to the rest of the program.