Modified compiler to only depend on primordial one
The extended Lispkit Lisp compiler (compiler.lso) was modified so that it may
be compiled using APENDIX2.LOB (a direct transcription from the Henderson
book), rather than needing a precompiled version of the compiler in order to
compile it (with the obvious chicken and egg problem that that entrails).

The Makefile was modified so that compiler.lob is built from APENDIX2.LOB (the
"primordial" compiler), using the following steps:

  1) The primordial compiler builds itself --> "primitive compiler"
  2) The primitive compiler (from step 1) builds the extended compiler.
  3) The extended compiler rebuilds itself.

As a result, compiler.lob is no longer needed in the repository and has been
removed.