Fixed issue with clobbered registers.
Reserved SECD-machine registers need to be preserved across an external
procedure call.  This affected the opcodes that read and write to the terminal
(PEXP, PUT, GET).

Also fixed a few other minor bugs (added a missing jump statement that should be
there but that the lack of it would not cause any harm), and added a NOP
opcode (op. 0) and corresponding LispKit command for diagnostic purposes.