Added BEGIN expressions to LispKit lisp.
The expression (BEGIN e₁ e₂ ... eₙ) evaluates the expressions e₁..eₙ in
sequential order and evaluates to eₙ.  A new opcode (POP) was added to the
SECD-machine which pops an item off the stack.   The compiler inserts POP
instructions after e₁, ..., eₙ₋₁, so that the stack will not grow with unused
results as these expressions are processed.