Modified existing LispKit code to use BEGIN.
The previous approach was to create a function FIRSTARG that took two args
and evaluated to its first argument.  The second argument would be evaluated
as a side effect prior to evaluating the first argument.  The problem with
this approach is that, although arguments are evaluated in reverse order, this
is not a required behavior and could change.  Now that a BEGIN statement has
been added for this purpose and the arguments are processed in a defined order,
we now use this instead.