Fixed LEQ opcode to compare types first.
LEQ should compare types, and only compare values if their types are equal.
This ensures that a symbol will not be considered equal to a number (i.e.,
(AND (LEQ A B) (LEQ B A)) evaluates to true) simply because the number happens
to be the address of the symbol's string in the string store.