Modified mark-phase of GC to use tail recursion.
The recursive call to _mark on the CDR of a cons-cell is in the tail position
(the "jmp .endif" instruction after the call is immediately followed by a "ret"
instruction).  So instead we now just jmp to the beginning of the function,
rather than using the "call" instruction, thereby preventing the stack from
growing with unnecessary instruction pointer (IP) values.  Now the mark
algorithm grows the stack in proportion to the maximum nesting depth of the
lists it is tracing, rather than in proportion to the maximum length of those
lists.