Add Landin's J-operator
For encapsulating the state of the SECD machine as a closure for later
invocation -- used to implement Scheme's "call/cc" operator.

See:

P.J. Landin. A generalization of jumps and labels. Report, UNIVAC
Systems Programming Research, August 1965.

H. Thielecke. An introduction to Landin's "A generalization of jumps
and labels", Higher-Order and Symbolic Computation 11(2):117-123, 1998