Replaced (F . ARGS) syntax with (APPLY F ARGS)
(F . ARGS) may be ambiguous if ARGS is not a variable reference, but a
function call.  (F . (G X)) looks like it means "evaluate G with the
argument X, and map the elements of the resulting list to the arguments
of F".  However, the S-expression "(F . (G X))" is equivalent to
"(F G X)", so what "(F . (G X))" really means is "evaluate F with the
arguments G and X".